Split IntroCarouselView file
